Pull handle_known_peers_events out into a new 'events' module
